feat(backend): add session auth for the api

+++ backend/src/controllers/expense.controller.ts
@@ -1,51 +1,75 @@
 import { NextFunction, Request, Response} from "express"
+import { ExtRequest } from "../interfaces"
+import { UserModel } from "../models"
 import { ExpenseModel } from "../models/expense.model"
 import { ExpenseService } from "../services"
+import { updateExpense } from "../services/expense.service"
 
-export const getAll = async (req: Request,res: Response, next: NextFunction) => {
+export const getAll = async (req: ExtRequest,res: Response, next: NextFunction) => {
   try {
-    const expenses = await ExpenseService.getAllExpenses()
+    const expenses = await ExpenseService.getAllExpenses(`${req.user?.id}`)
     res.status(200).json({expenses})
+    next()
   } catch (err) {
     next('ERROR_EXPENSE_GET_ALL')
   }
 }
 
-export const postOne = async (req: Request,res: Response, next: NextFunction) => {
+export const postOne = async (req: ExtRequest,res: Response, next: NextFunction) => {
   try {
     const { body } = req
-    const expense = await ExpenseService.createExpense(body)
+    const data = {
+      ...body,
+      userId: req.user?.id
+    }
+    const expense = await ExpenseService.createExpense(data)
     res.status(201).json({expense})   
   } catch (error) {
     next('ERROR_EXPENSE_POST_ONE')
   }
 }
 
-export const getOne = async ({params}: Request,res: Response, next: NextFunction) => {
+export const getOne = async (req: ExtRequest,res: Response, next: NextFunction) => {
   try {
-    const { id } = params
+    const { id } = req.params
     const expense = await ExpenseService.getExpenseById(id)
-    res.status(200).json({expense})
+    if(expense && expense.userId === req.user?.id){
+      res.status(200).json({expense})
+    }else{
+      next('EXPENSE_NOT_FOUND')
+    }
+    
   } catch (error) {
-    next('ERROR_EXPENSE_GET_ONE')
+    next('EXPENSE_NOT_FOUND')
   }
 }
 
-export const updateOne = async ({params,body}: Request,res: Response, next: NextFunction) => {
+export const updateOne = async (req: ExtRequest,res: Response, next: NextFunction) => {
   try {
-    const { id } = params
-    const expense = await ExpenseService.updateExpense(id,body)
-    res.status(200).json({expense})
+    const { id } = req.params
+    const expenseFound = await ExpenseService.getExpenseById(id)
+    if(expenseFound && expenseFound.userId === req.user?.id){
+      const updatedExpense = await ExpenseService.updateExpense(id,req.body)
+      res.status(200).json({expense: updatedExpense})
+    }else{
+      next('EXPENSE_NOT_FOUND')
+    }
+    
   } catch (error) {
     next('ERROR_EXPENSE_UPDATE_ONE')
   }
 }
 
-export const removeOne = async ({params}: Request,res: Response, next: NextFunction) => {
+export const removeOne = async (req: ExtRequest, res: Response, next: NextFunction) => {
   try {
-    const { id } = params
-    const expense = await ExpenseService.removeExpense(id)
-    res.status(200).json({expense})
+    const { id } = req.params
+    const expenseFound = await ExpenseService.getExpenseById(id)
+    if(expenseFound && expenseFound.userId === req.user?.id){
+      const expense = await ExpenseService.removeExpense(id)
+      res.status(200).json({expense})
+    }else{
+      next('EXPENSE_NOT_FOUND')
+    }
   } catch (error) {
     next('ERROR_EXPENSE_REMOVE_ONE')
   }
